+++ services/API-service/src/api/admin-area-dynamic-data/enum/dynamic-indicator.enum.ts
